A minimal webserver that serves the value of a variable
This is for the lessons on Javascript and node.js at SocialHackersAcademy (https://www.socialhackersacademy.org/)

First version of the exercise.
This is for week 1 of module NodeJS.

we just expose the value of a variable. We can see the value, reset it, add to it or subtract from it.